+++ packages/li-editor/src/widgets/TilesetsDataset/types.ts
@@ -3,3 +3,10 @@ export type XYZTilesetConfig = {
   url: string;
   zoom: { minZoom?: number; maxZoom?: number };
 };
+
+export type MVTTilesetConfig = {
+  name: string;
+  url: string;
+  metadataUrl: string;
+  zoom: { minZoom?: number; maxZoom?: number };
+};
